Fixed debug build (format string/argument mismatches).
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@38619 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/kits/tracker/DeskWindow.cpp
@@ -195,7 +195,7 @@ BDeskWindow::MenusBeginning()
 		// remove all current addon shortcuts
 		for (std::set<uint32>::iterator it= fCurrentAddonShortcuts.begin();
 			it != fCurrentAddonShortcuts.end(); it++) {
-			PRINT(("removing shortcut %c\n", *it));
+			PRINT(("removing shortcut %c\n", (int)*it));
 			RemoveShortcut(*it, B_OPTION_KEY | B_COMMAND_KEY);
 		}
